initial commit
Merged an API parser (ApiTools), dex modding tools
(DexTools, HookTool, MergeTool), an APK modding tools
(ApkTools), and an AndroidManifest.xml modding tool
(ManifestTools) into this project.

HookTool and ApiTools should work together. They don't
at the moment.

There should be an option to reuse signing keys. Likewise,
there should be an option to select the signing key.

This is using an old version of dexlib. This may or may
not be a problem, depending on the target Android version.

There's pretty much no documentation. This gets bumped up
in priority if anyone else actually starts deving on this.